Changed len types to long

+++ shared/src/main/kotlin/dev/luna5ama/glwrapper/impl/BufferObject.kt
@@ -6,14 +6,14 @@ import dev.luna5ama.kmogus.Ptr
 sealed class BufferObject : IGLObject, IGLTargetBinding {
     override val id = glCreateBuffers()
 
-    var size = 0; private set
+    var size = 0L; private set
 
-    open fun allocate(size: Int, flags: Int): BufferObject {
+    open fun allocate(size: Long, flags: Int): BufferObject {
         this.size = size
         return this
     }
 
-    open fun allocate(size: Int, data: Ptr, flags: Int): BufferObject {
+    open fun allocate(size: Long, data: Ptr, flags: Int): BufferObject {
         this.size = size
         return this
     }
@@ -42,29 +42,29 @@ sealed class BufferObject : IGLObject, IGLTargetBinding {
     }
 
     class Immutable : BufferObject() {
-        override fun allocate(size: Int, flags: Int): Immutable {
+        override fun allocate(size: Long, flags: Int): Immutable {
             super.allocate(size, flags)
-            glNamedBufferStorage(id, size.toLong(), Ptr.NULL, flags)
+            glNamedBufferStorage(id, size, Ptr.NULL, flags)
             return this
         }
 
-        override fun allocate(size: Int, data: Ptr, flags: Int): Immutable {
+        override fun allocate(size: Long, data: Ptr, flags: Int): Immutable {
             super.allocate(size, data, flags)
-            glNamedBufferStorage(id, size.toLong(), data, flags)
+            glNamedBufferStorage(id, size, data, flags)
             return this
         }
     }
 
     class Mutable : BufferObject() {
-        override fun allocate(size: Int, flags: Int): Mutable {
+        override fun allocate(size: Long, flags: Int): Mutable {
             super.allocate(size, flags)
-            glNamedBufferData(id, size.toLong(), Ptr.NULL, flags)
+            glNamedBufferData(id, size, Ptr.NULL, flags)
             return this
         }
 
-        override fun allocate(size: Int, data: Ptr, flags: Int): Mutable {
+        override fun allocate(size: Long, data: Ptr, flags: Int): Mutable {
             super.allocate(size, data, flags)
-            glNamedBufferData(id, size.toLong(), data, flags)
+            glNamedBufferData(id, size, data, flags)
             return this
         }
     }
